refactor(File): explicitly import utils
