--- frappe/core/doctype/file/__init__.py
@@ -1,2 +1,0 @@
-from .exceptions import *
-from .utils import *
